fixing admin

+++ src/ReviewBundle/Controller/ReviewController.php
@@ -3,7 +3,9 @@
 namespace ReviewBundle\Controller;
 
 use AlbumBundle\Entity\Album;
-use TrackBundle\Entity\Track;
+use BlogBundle\Entity\Entry;
+use BlogBundle\Form\EntryType;
+use Exception;
 use Knp\Component\Pager\Paginator;
 use ReviewBundle\Entity\Review;
 use ReviewBundle\Form\AddReviewFormType;
@@ -12,8 +14,13 @@ use Symfony\Component\HttpFoundation\RedirectResponse as RedirectResponseAlias;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Security\Core\Exception\AccessDeniedException;
+use TrackBundle\Entity\Track;
 use UserBundle\Entity\User;
 
+/**
+ * Class ReviewController
+ * @package ReviewBundle\Controller
+ */
 class ReviewController extends Controller
 {
     /**
@@ -22,7 +29,7 @@ class ReviewController extends Controller
      * @param Request $request
      * @param int
      * @return Response
-     * @throws \Exception
+     * @throws Exception
      */
     public function addReviewAction(Request $request, $id)
     {
@@ -36,8 +43,7 @@ class ReviewController extends Controller
 
             if (!$album) {
                 $this->addFlash("error", "Failed to find Album!");
-            }
-            else {
+            } else {
 
                 try {
                     /** @var User $user */
@@ -45,14 +51,17 @@ class ReviewController extends Controller
                     $review->setReviewer($user);
                     /** @var Album $album */
                     $review->setAlbum($album);
-                    $review->setTitle($album->getTitle());
                     $review->setTimestamp(new \DateTime());
                     $em->persist($review);
                     $em->flush();
 
                     $this->addFlash("success", "Thank you for your review.");
-                }
-                catch (\Exception $e) {
+                    return $this->redirect($this->generateUrl('view_review', [
+                            'id' => $review->getId()
+                        ]
+                    ));
+
+                } catch (Exception $e) {
                     $this->addFlash("error", "Failed to persist your review.");
                 }
             }
@@ -70,11 +79,12 @@ class ReviewController extends Controller
      * @param $id
      * @return Response
      */
-    public function viewByAlbumAction(Request $request, $id) {
+    public function viewByAlbumAction(Request $request, $id)
+    {
         $em = $this->getDoctrine()->getManager();
         $query = $em->getRepository(Review::class)
             ->getReviewsByAlbumID($id);
-        $album =  $em->getRepository(Album::class)
+        $album = $em->getRepository(Album::class)
             ->find($id);
 
         if (!$album) {
@@ -92,7 +102,7 @@ class ReviewController extends Controller
         $paginator = $this->get('knp_paginator');
         $reviews = $paginator->paginate(
             $query,
-            $request->query->getInt('page', 1),$this->getParameter('page_limit')
+            $request->query->getInt('page', 1), $this->getParameter('page_limit')
         );
 
         $totalReviews = 0;
@@ -110,10 +120,10 @@ class ReviewController extends Controller
         }
 
         return $this->render('@Review/Default/viewByAlbum.html.twig', [
-               'album' => $album,
-               'reviews' => $reviews,
-               'rating' => $albumRating
-           ]
+                'album' => $album,
+                'reviews' => $reviews,
+                'rating' => $albumRating
+            ]
         );
     }
 
@@ -131,12 +141,11 @@ class ReviewController extends Controller
             ->find($id);
 
         if (!$review) {
-            $this->addFlash('warning', 'Review not found!.');
+            $this->addFlash('warning', 'Review not found!');
+            return $this->redirect($this->generateUrl('album_homepage'));
         }
 
-        return $this->render(
-            'ReviewBundle:Default:viewReview.html.twig',
-            [
+        return $this->render('ReviewBundle:Default:viewReview.html.twig', [
                 'review' => $review
             ]
         );
@@ -150,7 +159,60 @@ class ReviewController extends Controller
      * @param $id
      * @return RedirectResponseAlias|Response
      */
-    public function editReviewAction(Request $request, $id){
+    public function editReviewAction(Request $request, $id)
+    {
+//        $entityManager = $this->getDoctrine()->getManager();
+//        $entry = $entityManager->getRepository(Review::class)->find($id);
+//
+//        // blocks all user from editing the review except any admin and the user
+//        // that the review belongs to
+//        if ($entry->getReviewer() !== $this->getUser() &&
+//            !$this->container->get('security.authorization_checker')
+//                ->isGranted('ROLE_ADMIN')
+//        ) {
+//            throw new AccessDeniedException();
+//        }
+//
+//
+////        $form = $this->createForm(
+////            Review::class,
+////            $entry,
+////            [
+////                'action' => $request->getUri()
+////            ]
+////
+////
+////        );
+//
+//
+////        $form = $this->createForm(AddReviewFormType::class, $entry);
+//
+//        $review = new Review();
+//        $form = $this->createForm(AddReviewFormType::class, $review);
+//
+//        $form->handleRequest($request);
+//
+//
+//        var_dump('here');
+//        if ($form->isValid()) {
+//            $entityManager->flush();
+//            return $this->redirect($this->generateUrl(
+//                'view_review',
+//                [
+//                    'id' => $entry->getId()
+//                ]
+//            ));
+//        }
+//
+//        return $this->render(
+//            'ReviewBundle:Default:edit.html.twig',
+//            [
+//                'form' => $form->createView(),
+//                'review' => $entry
+//            ]
+//        );
+//    }
+
 
         $em = $this->getDoctrine()->getManager();
         $review = $em->getRepository(Review::class)
@@ -160,29 +222,42 @@ class ReviewController extends Controller
             $this->addFlash('warning', 'Review not found!.');
             return $this->redirect($this->generateUrl('album_homepage'));
         }
-        if ($this->getUser() !== $review->getReviewer() && !$this->container->get('security.authorization_checker')
+
+//        /** @var User $user */
+//        $user = $this->getUser();
+
+
+        if ($review->getReviewer() !== $this->getUser() &&
+            !$this->container->get('security.authorization_checker')
                 ->isGranted('ROLE_ADMIN'))
         {
             $this->addFlash('error', 'You are not allowed to edit this review as you do not own it!');
             return $this->redirect($this->generateUrl('album_homepage'));
         }
+        else {
+            $form = $this->createForm(AddReviewFormType::class, $review, [
+                'action' => $request->getUri()
+            ]);
 
-        $form = $this->createForm(AddReviewFormType::class,  $review, [
-            'action' => $request->getUri()
-        ]);
+            $form->handleRequest($request);
 
-        $form->handleRequest($request);
+            if ($form->isSubmitted() && $form->isValid()) {
+                $em->flush();
 
-        if ($form->isValid()) {
-            $em->flush();
+                $this->addFlash("success", "The review was updated.");
+                return $this->redirect($this->generateUrl(
+                    'view_review',
+                    [
+                        'id' => $review->getId()
+                    ]
+                ));
+            }
 
-            $this->addFlash("success", "The review was updated.");
+            return $this->render('ReviewBundle:Default:edit.html.twig', [
+                'form' => $form->createView(),
+                'review' => $review
+            ]);
         }
-
-        return $this->render('ReviewBundle:Default:edit.html.twig', [
-            'form' => $form->createView(),
-            'review' => $review
-        ]);
     }
 
     /**
@@ -192,19 +267,26 @@ class ReviewController extends Controller
      *
      * @return RedirectResponseAlias
      */
-    public function deleteReviewAction($id) {
+    public function deleteReviewAction($id)
+    {
         $em = $this->getDoctrine()->getManager();
         try {
             $review = $em->getRepository(Review::class)
                 ->find($id);
 
-            $em->remove($review);
-            $em->flush();
+            if ($review->getReviewer() !== $this->getUser() &&
+                !$this->container->get('security.authorization_checker')
+                    ->isGranted('ROLE_ADMIN')) {
+                $this->addFlash('error', 'You are not allowed to edit this review as you do not own it!');
+                return $this->redirect($this->generateUrl('album_homepage'));
+            } else {
+                $em->remove($review);
+                $em->flush();
 
-            $this->addFlash('success', 'The review was deleted!');
-            return $this->redirect($this->generateUrl('album_homepage'));
-        }
-        catch (\Exception $e) {
+                $this->addFlash('success', 'The review was deleted!');
+                return $this->redirect($this->generateUrl('album_homepage'));
+            }
+        } catch (Exception $e) {
             $this->addFlash('error', 'Failed to delete review!');
             return $this->redirect($this->generateUrl('view_reviews_by_album', [
                 'id' => $id
